LUM-13 -add user module
